+++ lib/Model/Subject.php
@@ -104,6 +104,7 @@ class Subject
     /**
      * Create and add a new variant based on this subject.
      *
+     * @param array<string,mixed> $computedStats
      */
     public function createVariant(ParameterSet $parameterSet, int $revolutions, int $warmup, array $computedStats = []): Variant
     {
@@ -114,19 +115,27 @@ class Subject
             $warmup,
             $computedStats
         );
-        $this->variants[$parameterSet->getName()] = $variant;
+        $this->variants[] = $variant;
 
         return $variant;
     }
 
-    public function setVariant(Variant $variant): void
+    public function addVariant(Variant $variant): void
     {
         if ($variant->getSubject() !== $this) {
             throw new RuntimeException(
                 'Adding variant to subject to which it does not belong'
             );
         }
-        $this->variants[$variant->getParameterSet()->getName()] = $variant;
+        $this->variants[] = $variant;
+    }
+
+    /**
+     * @deprecated Use addVariant. To be removed in 2.0
+     */
+    public function setVariant(Variant $variant): void
+    {
+        $this->addVariant($variant);
     }
 
     /**
@@ -134,7 +143,7 @@ class Subject
      */
     public function getVariants(): array
     {
-        return $this->variants;
+        return array_values($this->variants);
     }
 
     /**
@@ -145,16 +154,25 @@ class Subject
         return $this->benchmark;
     }
 
+    /**
+     * @return string[]
+     */
     public function getGroups(): array
     {
         return $this->groups;
     }
 
+    /**
+     * @param string[] $groups
+     */
     public function inGroups(array $groups): bool
     {
         return 0 !== count(array_intersect($this->groups, $groups));
     }
 
+    /**
+     * @param string[] $groups
+     */
     public function setGroups(array $groups): void
     {
         $this->groups = $groups;
@@ -232,9 +250,30 @@ class Subject
         });
     }
 
+    /**
+     * Returns the _first_ variant that matches the given parameter set name.
+     * Note that there may be multiple variants with the same parameter set as
+     * they can also vary by the number of revs/iterations.
+     */
+    public function getVariantByParameterSetName(string $parameterSetName): ?Variant
+    {
+        foreach ($this->variants as $variant) {
+            if ($variant->getParameterSet()->getName() !== $parameterSetName) {
+                continue;
+            }
+
+            return $variant;
+        }
+
+        return null;
+    }
+
+    /**
+     * @deprecated use getVariantByParameterSetName. will be removed in 2.0
+     */
     public function getVariant(string $parameterSetName): ?Variant
     {
-        return $this->variants[$parameterSetName] ?? null;
+        return $this->getVariantByParameterSetName($parameterSetName);
     }
 
     public function setFormat(?string $format): void