+++ tests/Unit/Model/SubjectTest.php
@@ -12,37 +12,21 @@
 
 namespace PhpBench\Tests\Unit\Subject;
 
-use PhpBench\Model\Benchmark;
 use PhpBench\Model\ParameterSet;
-use PhpBench\Model\Subject;
 use PhpBench\Tests\TestCase;
+use PhpBench\Tests\Util\SubjectBuilder;
 
 class SubjectTest extends TestCase
 {
-    private $subject;
-    private $benchmark;
-
-    protected function setUp(): void
-    {
-        $this->benchmark = $this->prophesize(Benchmark::class);
-        $this->benchmark->getSubjects()->willReturn([]);
-        $this->subject = new Subject($this->benchmark->reveal(), 'subjectOne', 0);
-    }
-
-    /**
-     * It should say if it is in a set of groups.
-     */
     public function testInGroups(): void
     {
-        $this->subject->setGroups(['one', 'two', 'three']);
-        $result = $this->subject->inGroups(['five', 'two', 'six']);
-        $this->assertTrue($result);
+        $subject = SubjectBuilder::create('one')
+            ->withGroups(['one', 'two', 'three'])
+            ->build();
 
-        $result = $this->subject->inGroups(['eight', 'seven']);
-        $this->assertFalse($result);
-
-        $result = $this->subject->inGroups([]);
-        $this->assertFalse($result);
+        self::assertTrue($subject->inGroups(['five', 'two', 'six']));
+        self::assertFalse($subject->inGroups(['eight', 'seven']));
+        self::assertFalse($subject->inGroups([]));
     }
 
     /**
@@ -50,17 +34,47 @@ class SubjectTest extends TestCase
      */
     public function testCreateVariant(): void
     {
+        $subject = SubjectBuilder::create('one')
+            ->build();
+
         $parameterSet = ParameterSet::fromSerializedParameters('foo', []);
-        $variant = $this->subject->createVariant(
+        $variant = $subject->createVariant(
             $parameterSet,
             10,
             20
         );
 
-        $this->assertEquals($this->subject, $variant->getSubject());
-        $this->assertEquals($parameterSet, $variant->getParameterSet());
-        $this->assertEquals(0, $variant->count());
-        $this->assertEquals(10, $variant->getRevolutions());
-        $this->assertEquals(20, $variant->getWarmup());
+        self::assertEquals($subject, $variant->getSubject());
+        self::assertEquals($parameterSet, $variant->getParameterSet());
+        self::assertEquals(0, $variant->count());
+        self::assertEquals(10, $variant->getRevolutions());
+        self::assertEquals(20, $variant->getWarmup());
+    }
+
+    public function testCreateMultipleVariantsWithSameParameterSetName(): void
+    {
+        $subject = SubjectBuilder::create('one')
+            ->build();
+
+        $parameterSet = ParameterSet::fromSerializedParameters('foo', []);
+        $variant = $subject->createVariant($parameterSet, 10, 20);
+        $parameterSet = ParameterSet::fromSerializedParameters('foo', []);
+        $variant = $subject->createVariant($parameterSet, 10, 20);
+
+        self::assertCount(2, $subject->getVariants());
+    }
+
+    public function testAddMultipleVariantsWithSameParameterSetName(): void
+    {
+        $subject = SubjectBuilder::create('one')
+            ->variant()
+                ->withParameterSet('foo', [])
+            ->end()
+            ->variant()
+                ->withParameterSet('foo', [])
+            ->end()
+            ->build();
+
+        self::assertCount(2, $subject->getVariants());
     }
 }