+++ tests/Util/SubjectBuilder.php
@@ -25,6 +25,11 @@ final class SubjectBuilder
      */
     private $benchmarkBuilder;
 
+    /**
+     * @var string[]
+     */
+    private $groups = [];
+
     public function __construct(?BenchmarkBuilder $benchmarkBuilder, string $name)
     {
         $this->name = $name;
@@ -41,7 +46,7 @@ final class SubjectBuilder
         return new self($builder, $name);
     }
 
-    public function variant(string $name): VariantBuilder
+    public function variant(?string $name = null): VariantBuilder
     {
         $builder = VariantBuilder::forSubjectBuilder($this, $name);
         $this->variantBuilders[] = $builder;
@@ -49,7 +54,7 @@ final class SubjectBuilder
         return $builder;
     }
 
-    public function build(?Benchmark $benchmark): Subject
+    public function build(?Benchmark $benchmark = null): Subject
     {
         if (null === $benchmark) {
             $suite = new Suite(
@@ -60,12 +65,23 @@ final class SubjectBuilder
         }
 
         $subject = new Subject($benchmark, $this->name);
+        $subject->setGroups($this->groups);
 
         foreach ($this->variantBuilders as $builder) {
             $subject->setVariant($builder->build($subject));
         }
 
         return $subject;
+    }
+
+    /**
+     * @param string[] $groups
+     */
+    public function withGroups(array $groups): SubjectBuilder
+    {
+        $this->groups = $groups;
+
+        return $this;
     }
 
     public function end(): BenchmarkBuilder