+++ tests/Util/SuiteBuilder.php
@@ -17,6 +17,11 @@ final class SuiteBuilder
      */
     private $benchmarkBuilders = [];
 
+    /**
+     * @var DateTime
+     */
+    private $date = null;
+
     public function __construct(string $name)
     {
         $this->name = $name;
@@ -25,6 +30,13 @@ final class SuiteBuilder
     public static function create(string $name): self
     {
         return new self($name);
+    }
+
+    public function withDateString(string $date): SuiteBuilder
+    {
+        $this->date = new DateTime($date);
+
+        return $this;
     }
 
     public function benchmark(string $name): BenchmarkBuilder
@@ -39,7 +51,7 @@ final class SuiteBuilder
     {
         $suite = new Suite(
             $this->name,
-            new DateTime()
+            $this->date ?? new DateTime()
         );
 
         foreach ($this->benchmarkBuilders as $builder) {